Universe-A/requirements.txt
PyQt6==6.6.1
requests==2.31.0
transformers==4.36.2
torch==2.2.0
Flask==3.0.2
Flask-SocketIO==5.3.6
prometheus-client==0.19.0
psutil==5.9.8
python-jwt==4.1.0
Werkzeug==3.0.1
langchain==0.1.0
python-dotenv==1.0.0
huggingface-hub==0.20.1
